NOUPSTREAM: iommu: force IOMMU passthrough mode for Ampere eMAG system
This is a work around for the Ampere eMAG system, current Linux does
not support IORT DMA aperture restriction. The Ampere eMAG system
has a 42 bit bus feeding into an IOMMU which is pegged to 48 bits.
In translation mode, this would cause incorrect IOVA allocation.
This is a temporary work around to force Ampere eMAG to passthrough
mode in the host kernel.

For RHEL-7.6, the implication of passthrough mode as default is
removed. Due to incorrect IOVA allocation, some NICs (like Intel i210,
Mellanox ConnectX-4) do not work and cause crashing at boot. This
patch forces to passthrough and helps system to boot.

Ampere CentOS 7.5 needs this patch to prevent system booting with
translation mode until supporting is done.

Change-Id: I883c72d62608645a203a7f4d91a5c8f1e931d566
Signed-off-by: Feng Kan <fkan@amperecomputing.com>
Signed-off-by: Toan Le <toanle@amperecomputing.com>
Signed-off-by: Khuong Dinh <khuong.dinh@amperecomputing.com>